perf: prefetch 하는 layout 레벨 개선 및 HomePage 파일 이동 (#266)
